Backend/requirements.txt
Flask
pymongo
flask-bcrypt
flask-jwt-extended
python-dotenv
flask-pymongo
python-dotenv